Refactor and add ResilientWebSocket
